feat:CRM合同创建

+++ src/main/java/com/bdaim/crm/erp/crm/entity/base/BaseCrmContract.java
@@ -68,20 +68,20 @@ public abstract class BaseCrmContract<M extends BaseCrmContract<M>> extends Mode
 		return get("order_date");
 	}
 
-	public void setCreateUserId(Integer createUserId) {
+	public void setCreateUserId(Long createUserId) {
 		set("create_user_id", createUserId);
 	}
 
-	public Integer getCreateUserId() {
-		return getInt("create_user_id");
+	public Long getCreateUserId() {
+		return getLong("create_user_id");
 	}
 
-	public void setOwnerUserId(Integer ownerUserId) {
+	public void setOwnerUserId(Long ownerUserId) {
 		set("owner_user_id", ownerUserId);
 	}
 
-	public Integer getOwnerUserId() {
-		return getInt("owner_user_id");
+	public Long getOwnerUserId() {
+		return getLong("owner_user_id");
 	}
 
 	public void setCreateTime(Timestamp createTime) {
